hiding Javers constructor
